Remove redundant base64 encoding for private_key_base64 parameter.
The PKCS 8 file already provides base64 encoded string of the private
key, no need for extra base64 encode/decode. This slows dows the
connection with the database.

This PR removes extra base64 decode/encode step, improving the peformance
of private key parsing for "private_key_base64" parameter. Notice, that
the parameter is still "spark" serializable as it is a java String.

Unfortunately not backward compatible. If accepted by the maintainers
I would be happy to create a new property: "private_key_pkcs8_pem" to make
this change backward compatible.